feat(TCOMP-568): partial revert of junits and refactor to avoid cyclic dependencies for the release Revert "feat(TCOMP-568) add service tracker (#821)" This reverts commit 4c49e1cd8bfe0c84ea956fdffb68121a0804900c. Only partial revert, the commit https://github.com/Talend/components/pull/821/commits/74f89407000a6e128d1e75f250bf761ca4b73be3 was kept
