--- core/components-osgi-test/osgi-test-component/pom.xml
@@ -1,94 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-
-	<parent>
-		<groupId>org.talend.components</groupId>
-		<artifactId>components-parent</artifactId>
-		<version>0.19.0-SNAPSHOT</version>
-		<relativePath>../../../components-parent/pom.xml</relativePath>
-	</parent>
-
-	<artifactId>osgi-test-component</artifactId>
-	<version>0.19.0-SNAPSHOT</version>
-	<name>Components - TestComponent Definition</name>
-	<packaging>bundle</packaging>
-	<repositories>
-		<repository>
-			<id>talend_nexus</id>
-			<name>snapshots</name>
-			<url>https://artifacts-oss.talend.com/nexus/content/repositories/TalendOpenSourceSnapshot/</url>
-		</repository>
-		<repository>
-			<id>releases</id>
-			<url>https://artifacts-oss.talend.com/nexus/content/repositories/TalendOpenSourceRelease/</url>
-		</repository>
-	</repositories>
-
-	<dependencies>
-
-		<!-- APIs for all Components -->
-		<dependency>
-			<groupId>org.talend.components</groupId>
-			<artifactId>components-api</artifactId>
-		</dependency>
-
-		<!-- Annotation processing -->
-		<dependency>
-			<groupId>biz.aQute.bnd</groupId>
-			<artifactId>annotation</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>com.google.auto.service</groupId>
-			<artifactId>auto-service</artifactId>
-		</dependency>
-		<!-- Shared component properties such as Use/Password or Schema -->
-		<dependency>
-			<groupId>org.talend.components</groupId>
-			<artifactId>components-common</artifactId>
-			<version>0.16.0</version>
-		</dependency>
-
-
-	</dependencies>
-	<build>
-		<plugins>
-			<plugin>
-				<groupId>org.apache.felix</groupId>
-				<artifactId>maven-bundle-plugin</artifactId>
-				<extensions>true</extensions>
-				<executions>
-					<execution>
-						<id>osgi-bundle</id>
-						<phase>package</phase>
-						<goals>
-							<goal>bundle</goal>
-						</goals>
-						<configuration>
-							<instructions>
-								<Bundle-SymbolicName>org.talend.components.table</Bundle-SymbolicName>
-								<Service-Component>*</Service-Component>
-								<Embed-Dependency>*;scope=compile|runtime;inline=false;artifactId=!components*|auto-service|pax-url-aether</Embed-Dependency>
-								<Include-Resource>{maven-resources},META-INF/services/=target/classes/META-INF/services/,META-INF/maven/${project.groupId}/${project.artifactId}/dependencies.txt=target/classes/META-INF/maven/${project.groupId}/${project.artifactId}/dependencies.txt</Include-Resource>
-								<Bundle-ActivationPolicy>lazy</Bundle-ActivationPolicy>
-							</instructions>
-						</configuration>
-					</execution>
-				</executions>
-			</plugin>
-			<plugin>
-				<artifactId>maven-source-plugin</artifactId>
-				<configuration>
-					<skipSource>true</skipSource>
-				</configuration>
-			</plugin>
-			<plugin>
-				<artifactId>maven-javadoc-plugin</artifactId>
-				<configuration>
-					<skip>true</skip>
-				</configuration>
-			</plugin>
-		</plugins>
-	</build>
-</project>
