--- core/components-osgi-test/osgi-test-component/src/main/java/org/talend/components/testcomponent/TestComponentFamilyDefinition.java
@@ -1,42 +1,0 @@
-// ============================================================================
-//
-// Copyright (C) 2006-2017 Talend Inc. - www.talend.com
-//
-// This source code is available under agreement available at
-// %InstallDIR%\features\org.talend.rcp.branding.%PRODUCTNAME%\%PRODUCTNAME%license.txt
-//
-// You should have received a copy of the agreement
-// along with this program; if not, write to Talend SA
-// 9 rue Pages 92150 Suresnes, France
-//
-// ============================================================================
-package org.talend.components.testcomponent;
-
-import org.talend.components.api.AbstractComponentFamilyDefinition;
-import org.talend.components.api.ComponentInstaller;
-import org.talend.components.api.Constants;
-import org.talend.components.testcomponent.testcomp.TestComponentDefinition;
-
-import aQute.bnd.annotation.component.Component;
-
-import com.google.auto.service.AutoService;
-
-/**
- * Install all of the definitions provided for the TestComponent family of components.
- */
-@AutoService(ComponentInstaller.class)
-@Component(name = Constants.COMPONENT_INSTALLER_PREFIX + TestComponentFamilyDefinition.NAME, provide = ComponentInstaller.class)
-public class TestComponentFamilyDefinition extends AbstractComponentFamilyDefinition implements ComponentInstaller {
-
-    public static final String NAME = "TestComponent";
-
-    public TestComponentFamilyDefinition() {
-        super(NAME, new TestComponentDefinition());
-
-    }
-
-    @Override
-    public void install(ComponentFrameworkContext ctx) {
-        ctx.registerComponentFamilyDefinition(this);
-    }
-}
