--- core/components-osgi-test/pom.xml
@@ -1,21 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <groupId>org.talend.components</groupId>
-    <artifactId>components-osgi-test-aggregator</artifactId>
-    <version>0.19.0-SNAPSHOT</version>
-
-    <name>Components - Osgi test Aggregator</name>
-    <packaging>pom</packaging>
-
-    <properties>
-        <!-- skip the deployment for this aggregator, not necessary -->
-        <maven.deploy.skip>true</maven.deploy.skip>
-    </properties>
-
-    <modules>
-        <module>components-osgitest</module>
-        <module>osgi-test-component</module>
-    </modules>
-
-</project>
