+++ core/components-osgitest/pom.xml
@@ -1,12 +1,11 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 	<modelVersion>4.0.0</modelVersion>
 
 	<parent>
 		<groupId>org.talend.components</groupId>
 		<artifactId>components-parent</artifactId>
 		<version>0.19.0-SNAPSHOT</version>
-		<relativePath>../../../components-parent/pom.xml</relativePath>
+		<relativePath>../../components-parent/pom.xml</relativePath>
 	</parent>
 
 	<properties>
@@ -17,12 +16,6 @@
 	<artifactId>components-osgitest</artifactId>
 
 	<dependencies>
-		<dependency>
-			<groupId>org.talend.components</groupId>
-			<artifactId>osgi-test-component</artifactId>
-			<version>${project.version}</version>
-			<scope>test</scope>
-		</dependency>
 		<dependency>
 			<groupId>org.talend.components</groupId>
 			<artifactId>components-api-service-osgi</artifactId>
@@ -39,11 +32,6 @@
 			<artifactId>components-api-service-common</artifactId>
 			<classifier>tests</classifier>
 			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>org.ops4j.pax.url</groupId>
-			<artifactId>pax-url-link</artifactId>
-			<version>2.4.7</version>
 		</dependency>
 		<dependency>
 			<groupId>org.talend.components</groupId>
@@ -152,7 +140,6 @@
 				<version>0.4.2-Talend</version>
 				<executions>
 					<execution>
-						<id>generated-link-files</id>
 						<goals>
 							<goal>generate-resources</goal>
 						</goals>
@@ -161,22 +148,6 @@
 							<value>mvn:%groupId%/%artifactId%/%version%//%classifier%</value>
 							<!-- using build directory to launch test from eclipse -->
 							<outputDirectory>${project.build.outputDirectory}</outputDirectory>
-							<separator>${line.separator}</separator>
-							<dependencySet>
-								<scope>test</scope>
-							</dependencySet>
-						</configuration>
-					</execution>
-					<execution>
-						<id>generated-link-files-for-current-tests</id>
-						<goals>
-							<goal>generate-resources</goal>
-						</goals>
-						<configuration>
-							<name>%groupId%-%artifactId%%dashClassifier%.link</name>
-							<value>mvn:%groupId%/%artifactId%/%version%//%classifier%</value>
-							<!-- using build directory to launch test from eclipse -->
-							<outputDirectory>${project.build.directory}/test-classes</outputDirectory>
 							<separator>${line.separator}</separator>
 							<dependencySet>
 								<scope>test</scope>
