+++ core/components-osgitest/src/main/java/org/talend/components/api/ComponentsPaxExamOptions.java
