+++ core/components-osgitest/src/test/java/org/talend/components/api/ComponentsPaxExamOptionsTest.java
